Fixed several errors with calls and tests.
The inclusive_scan_operations.py file was not calling the correct function, had to add that in with the correct c pointer calls.
Dove into the c++ files for the dim assertion error, realized that was just an issue with the c++ file, fixed that as well.
The test_inclusive.py file didn't have F16 coverage, checked for that as well.